Revert "Refactor/file structure organization"
