+++ src/components/_currencycard.scss
