+++ src/components/_error.scss
