+++ src/components/_main.scss
