+++ src/components/_popup.scss
