+++ src/index.scss
@@ -1,4 +1,4 @@
 @import 'vendor/normalize/import-now';
 @import 'vendor/fonts/index.scss';
-@import 'style/variables/index.scss';
-@import 'style/_index.scss';
+@import 'variables/index.scss';
+@import 'components';
